requirements.txt
# Core ML
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.0.0
imbalanced-learn>=0.12.0

# Explainability
shap>=0.45.0

# Experiment Tracking
mlflow>=2.14.0

# RAG & LLM
langchain>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0
langchain-text-splitters>=0.2.0
faiss-cpu>=1.7.0
sentence-transformers>=3.0.0

# API
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.0

# Dashboard
streamlit>=1.35.0
plotly>=5.20.0

# Data & Utils
requests>=2.32.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
loguru>=0.7.0

# Monitoring
evidently>=0.4.0

# Testing
pytest>=8.0.0
httpx>=0.27.0
